requirements.txt
tensorflow
keras
numpy
pandas
fastapi
uvicorn
gunicorn
jupyter
keras_nlp
cassandra-driver
pyasyncore
